feat: add STRING_CHARACTER RegExp
Although it is very horrid. I will probably come up with a DSL that
allows me to built it up out of the constituent chunks:

    alternate(
      ESCAPED_CHARACTER,
      ESCAPED_UNICODE,
      exclude(['\n', '\r', '"', '\\'], SOURCE_CHARACTER)
    )

That last one, `exclude()` is the only tricky one to write. It's pretty
easy to do if you know that SOURCE_CHARACTER corresponds to a character
class, as it does here, but if it's an arbitrary RegExp then I don't see
a trivial way to do it off the top of my head (although there is
probably some well-known automata algorithm to do it).