test: add STRING_VALUE demo
Although the resulting machine is pretty obviously wrong.

If we look at the STRING_CHARACTER transition table, it looks pretty
right to me, the salient points being:

- Move from state 0 to 1 for most characters (ie. accept).
- For a backslash, move to state 2, and you're allowed to follow that by
  a number of legit escape sequence characters (t, r, n, f, b, \, ", u)
  and only one I'm not sure about: / (but on checking the spec, it is
  legit).
- Seeing a "u" takes you to states 3, 4, 5, 6, which are used to
  recognize hex digits.

Compare that to STRING_VALUE:

- From 0 to 1 on: ", opening the string.
- From 1 to 2 on: ", accepting the string (eg. "" is valid).
- From 1 back to 0 on: \, meaning that the only character allowed to follow a
  backslash is a ".
- From 1 stay in 1 on: Tab, printable ASCII, a most or SOURCE_CHARACTER
  (up to \uffff).

So this machine will accept stuff like:

  ""
  "foo"
  "\""

But it won't accept other legit backslash escapes like:

  "hey\nthere"
  "this: \u0020"

My next step is going to be to dig into the intermediate steps and see
where I am discarding some of the necessary behavior.