feat: add first bits of IntervalTree
We'll be using this to avoid quadratic (or worse) runtime comparing
every condition against every other condition.

Simple example: "A-Z" going to state 1 and "J" going to state 2 need
to be split into "A-I" going to 1, "J" going to 2, and "K-Z" going to 1.
Note that as we split overlapping states like that, the number of
conditions will actually grow, which is why I say worse than quadratic.

Anyway, this is what I am eventually going to implement:

    https://en.wikipedia.org/wiki/Interval_tree

but there are a lot of combos and edge cases to handle, which is why I
am making these notes first.

My NFAs are small, so I could probably brute force this, but I don't
think that the brute force code will be much simpler than the code that
uses the more sophisticated data structures, so I may as well go for it.