feat: teach Red-Black Tree about deletion
I have some bugs in here still, as evidenced by the commented-out tests,
but it mostly works, I think.

This is very directly based on:

-  https://www.cs.princeton.edu/~rs/talks/LLRB/LLRB.pdf
-  https://www.cs.princeton.edu/~rs/talks/LLRB/RedBlack.pdf
-  https://www.cs.princeton.edu/~rs/talks/LLRB/Java/RedBlackBST.java
-  https://algs4.cs.princeton.edu/33balanced/RedBlackBST.java.html

There are some subtle discrepancies between these, so I'm trying to
identify the most robust/complete combination and "prove" that it's
valid via the tests. Will be adding more tests soon, as well as fixing
the existing bugs.