refactor: use `join()` instead of template literal interpolation
The wrapping on the interpolation looks ugly, even though it is
probably closer to my intent (ie. "print this, then a space, then
this"). Get rid of the wrapping by using a `join()`. This is more
readable, even though the intent isn't quite as simple/imperative (eg.
it's "join this and that using a space").