fix: improve rendering of variable-width keys
The problem here was that the code that draws the subtrees wanted the
subtree boxes to be as close together as possible (ie. "centered"
alignment), and the code that drew the edges assumed the boxes were as
far apart as possible (ie. "justified" alignment).

So, I had the choice here of:

- Making the edge code more complicated, and drawing narrower
  connections when appropriate (ie. "centered"); or:
- Making the subtree box code mode complicated, and drawing the boxes as
  far apart as possible (ie. "justified").

In the end, the additional complexity doesn't seem that great either way
(although I don't really want to be my future self if I have to come
back and understand how all this works 6 months from now), so I went
with the first option, because I think it produces nicer looking trees.

Can we make this simpler? I can't see a way to, so we'll have to live
with the complexity for now at least...