fix: make deletion work reliably in RedBlackTree
Lots of changes here that I used to help me zero in on the bug (eg.
integrity checks), but the root cause was my lame `_flipColors()`
implementation, which worked for flipping from

       R                              B
      / \             to             / \
     B   B                          R   R

but not the other way around. This is what I get for coding late at
night and when feeling under the weather.