fix: fix failure to traverse to right subtree in IntervalTree
This error comes right out of the Wikipedia page:

    https://en.wikipedia.org/wiki/Interval_tree#Java_example:_Searching_a_point_or_an_interval_in_the_tree

Either that, or I misunderstood the code as I was translating it from
Java into TS.

Whatever the case, we weren't traversing the right subtree when we
should. I caught the bug red-handed while looking at a tree with this
basic structure:

      [c,e]:f
          \
          [f]:f

It wasn't traversing the right link because `interval.low` (a) was less
than `h.key.low` (c), preventing it from seeing the overlapping `f`.

The check needs to be changes such that we avoid going down the right
subtree if and only if it can't possibly contain a match. This happens
when the low value of the node is greater than the high value of the
range we're looking for (because we use the low values for ordering
within the BST). That is, in our example, we know that if we were
looking for `a-b` we would never find it to the right of `c-e`, but we
certainly might find `a-f`.