fix: make STRING_VALUE DFA work
Inspecting the test which was failing we could see the following:

1. We'd get to the point where we wanted to insert the interval `a,f`.
2. We'd correctly detect these overlaps:

   - 0x5d,a
   - b
   - c,e
   - f

3. So, we'd do this on our first run through the loop, considering only
   the first overlap:

   - Delete 0x5d,a
   - Insert 0x5d,`
   - Insert a
   - Insert b,f

4. On our second run through loop, considering the second overlap:

   - Delete b
   - Insert b,e

And as you can see, we're already doing it wrong by that point.
Visually, these are the ranges we're dealing with:

    overlaps: *********************|*|***|*
    interval:                     a|b|cde|f

So, as soon as we got to "b" we were inserting all the way to "b,f".
Instead, what we need to do is make sure that we only do a greedy
insertion like that if there are no more overlaps to process. A simple
adjustment to the loop suffices to take that into account.

The generated DFA looks optimal to me, and checks out when rendered with
`dot`.