refactor: consolidate edges with adjacent conditions
This gives us fewer edges, which is nice. I won't say much about this
because the comments in the code make it pretty obvious, and you can see
the effect in the updated test output.

Note that I wrote this the dumb/simple way, not bothering to check
whether there are any edges that can be merged (ie. I always blow away
and regenerate the `edges` property). Our machines are small enough in
this project that that potentially needless work is relatively cheap
anyway, and it avoids some complexity.