perf: use `make` to run `dot` in parallel
I could also do this in JS, but `make` is better, because we can
trivially get it to only re-run the commands when the input files
change, making it much faster.

A note about the use of `@` or not to make some commands silent: there
is rhyme and reason here — we don't foreshadow commands that are going
to produce revealing output anyway (eg. `@yarn build` doesn't need
foreshadowing, because it is going to run and print `tsc --build`
anyway) but we do do it for commands that would otherwise be silent (eg.
`rm $stuff`). A thing like the `@for` loop is silent because we trust
the recursive Makefiles to do their own output when appropriate.