style: don't format tsconfig-base.json
Prettier wants to mess with the whitespace, which makes comparing old
and new versions harder (even when viewing with `git diff -w`), so let's
just freeze it as-is.

I'm also removing the exclusion of `packages/compiler/src/Lexer.ts` as,
whatever the problem used to be, it doesn't seem to be an issue any
longer with the current version of Prettier (and besides, this file is
going away once the new lexer is finished, which will be Real Soon
Now™️).