chore: add first pass at BLOCK_STRING_VALUE
This is the naive approach, just to see if it will work, which consists
of making a single regexp and seeing if that captures the semantics.

It doesn't.

The built machine will recognize inputs like:

  """"""
  """hey there"""

But it will also accept stuff like this:

  """woah \"""
  """bad"""""""""""""""""""

The former via self transition from (3) back to (3) for "\", followed by
""" taking us to states (4), (5), and (6) which is an accept state. The
latter because once in (6) we have a self transition on '"' back to (6).

Not surprising because the regex is flawed; it matches the same things.
In other words, I think my regexp-to-DFA logic is probably fine. I had
been hoping that it might do a greedy match thing that would save it
from pursuing the wrong path here, but I don't think it works that way.

Going to try swapping the order of the alternate block next.