fix: add NUMBER pattern and make it work
This revealed a bug in the duplicate edge removal code in the
`removeEpsilons()` function. I fixed that, and also moved the check
outside of the loop. It was already horribly inefficient, but running it
every time through the loop made it even worse.

As seen in the diagram, the result DFA looks good. Although the GraphQL
spec goes into great detail distinguishing between floats and integers,
the resulting tokens both map onto JS numbers, so we can use them as-is
(in this project we don't have to worry about the vagaries of IEEE 754
floats; we only ever actually deal with simple numbers like "3", "10"
and so on).