chore: add BLOCK_STRING_VALUE_TT
So, I tried some different regexps and couldn't get them to do exactly
what I want (didn't commit them seeing as it wasn't clear that they
constituted forward progress, even though some of them did seem to
produce slightly better machines).

Here trying a different tack: this is a hand-written state machine that
I believe does what I want. Now I just have to figure out how to write
that as a regular expression (which is definitely possible due to the
equivalence of NFAs and regexps), and then confirm that that regexp
(which is probably horrendously ugly) doesn't get mangle when I plug it
back in through the regexp-to-minimized-DFA machinery.