fix: make sure we can scan NAME (etc) at end of input
This is a bit of a hack. Problem is that we could be in the middle of a
NAME token, and we get to the end of the input, but we don't know we're
there yet, so we continue to the top of the while loop to see if the
NAME continues, but of course it doesn't, because we don't even enter
the loop body and we therefore never yield the token.

The hack is a hack because the last call to `input.charCodeAt(i)` is
actually going to read beyond the end of the input, so it is going to
return `NaN`. That is enough to have the desired effect. Specifically,
we drop into state `18` one last time:

    case 18:
      if (
        (ch >= 0x30 && ch <= 0x39) ||
        (ch >= 0x41 && ch <= 0x5a) ||
        ch === 0x5f ||
        (ch >= 0x61 && ch <= 0x7a)
      ) {
        state = 18;
      } else {
        yield {
          token: 'NAME',
          tokenStart,
          tokenEnd: i,
        };
        tokenStart = i;
        state = START;
        continue loop;
      }
      break;

and because `ch` is `NaN`, we wind up in the `else`. All of those
comparisons are going to be false because:

    NaN >= 0x30 // false
    NaN <= 0x39 // false
    NaN === 0x5f // false

So, in the `else` we yield the token, and we're done. I'm actually
confused as to why this isn't an infinite loop, TBH, because we're not
incrementing `i`. But obviously something must explain why it's
working... 🤔