perf: inline inlineable DFA states
There are some accept states that:

1. Are reached from exactly one state.
2. Have no outbound edges.

Instead of transitioning to these states, which don't do anything except
emit tokens, we can just emit the token as soon as we see the edge that
would take us to those states, then eliminate the states.

This won't produce any perceptible perf boost (see below), but it is
strictly less work. As always, the cost is implementation complexity,
and possibly also the confusing fact that the compiled version of the
transition table doesn't actually have all the states that the table
had. So maybe this is a terrible idea, but I can't resist bad ideas. And
the truth is, my guiding principle here is that I want this code to
generate the state machine that I would have written by hand, and
tweaked by hand, and this is exactly the kind of micro-optimization that
I would have done by hand, so there you have it...

Numbers: I ran the benchmarks before and after about 10 times, but
there is no reliably discernible difference. Example:

- A bunch of runs with the lexers prior to this commit:

  47.072754
  44.254343
  44.760125
  32.990586
  44.11749
  44.634778
  44.277096
  44.710621
  44.539023
  45.450495
  44.209335
  32.905813
  44.942423
  45.011614

- A bunch of runs at this commit:

  45.607633
  33.202308
  44.158833
  44.303858
  44.16016
  43.606067
  44.204001
  44.559202
  45.668215
  44.880377
  45.054009
  44.25838

Thanks to Heartbleed (or whatever it was), the resolution of the Node
timer is pretty crappy anyway nowadays, so without increasing the
runtime significantly, I don't think we're going to get any signal here.