feat(parser): generate LALR item sets
This is not the full action/goto table yet, but it produces a necessary
intermediate artifact. At first glance, the output seen in the tests
seems about right. Basically, this is like a standard LR(1) item set
automaton, but with duplicate states consolidated; that is, anything
with a matching "core" (same rules, same LHS, same RHS, same dot
positions) is considered equal, and we just merge the lookahead symbols.

At the moment, I am not actually recording transitions anywhere, and I
am not trying to make this particularly performant. As you can see,
there are a lot of loops. But it doesn't need to be blazing fast; it
just needs to work for this grammar.