feat: show parser generator working for toy grammar
Using the grammar from:

    https://web.archive.org/web/20211216015406/https://web.cs.dal.ca/~sjackson/lalr1.html

Because that one has details I can use to verify that the various
substeps are doing the right thing; eg:

- Computing first and follow sets.
- Computing item sets.
- Computing a transition table.
- Computing a parse table.
- Actually executing the automata.
- etc

It all seems to be working, although with some obvious inefficiencies
that I don't care about right now (overall, I was optimizing for
readability and directness). Also, the plumbing is in there for running
actions during parsing, but I'm not actually doing it yet.