fix: eliminate TS errors
I had these commented out because during development in an intermediate
state I was returning invalid types. But now that the dynamic parser
is working, at least for the grammars in the test suite, seems we can
afford to bring back the assertions, which in turn eliminates the TS
errors. Notwithstanding that, I expect all of this to change in the
future.