refactor(lexer): switch tokens from POJOs to class instances
Going to need access to the contents for the parser (not much use
parsing a `NAME` if you can't easily get at the original text). We could
do this by passing the `input` into the parser and giving it some helper
function like:

    function getTokenText(token: Token, input: string): string { ... }

but it's going to feel a lot more ergonomic to just read
`token.contents` whenever needed.

I usually don't like classes, but:

- Most tokens don't need access to their text.
- We can therefore implement lazy getters at low cost by using class
  instances (only one instance of the method exists on the prototype,
  whereas we'd need to create multiple copies on POJOs, or create a
  shared copy and close over it, which would feel clumsy).

Fun fact: needed to add the `as const` to the `ast` variable because
otherwise TS mistakenly thinks there is no `new` property on it, even
though there obviously is (and TS's own error messaeg makes reference to
it):

    packages/lexer/src/build.ts:307:27 - error TS2339: Property 'new'
    does not exist on type '{ new (object: string | Expression,
    ...args: Expression[]): NewExpression; assign(binding:
    "const" | "let" | "var" | null, lhs: string, rhs: number |
    Expression): AssignmentStatement; ... 18 more ...; readonly
    true: BooleanValue; }'.

So, we can rename it (eg. `neww` works fine) or use the `as const` hack;
going with the latter.

You can also see in here some shenanigans involving `vm` so that I can
inject the new `Token` class into the "global" scope where I'm executing
the `lex` generator function. The old `GeneratorFunction` trick goes out
the window.