feat(parser): start building up parser
Just a demo of the kind of small step I'll be taking (probably the first
of about 100 such steps) in order to head towards complete coverage of
the grammar. From here, AST node types will get added and grown (ie. the
`FIELD` type is still incomplete, as it needs to handle optional
directives, arguments, and sub-selections), and more and more rules will
be added.