feat(parser): handle (some) epsilon productions
Prior to this commit,  the parser generator did not support epsilon
productions, so my tiny little grammar had this (partial) set of
rules for recognizing fields containing an optional `Alias` without
using epsilons:

    Field → NAME
    Field → Alias NAME
    Alias → NAME COLON

That's all well and good and seemed to work (I don't know if by blind
luck or because the generator is working correctly), but fields actually
have a lot of optional components:

    Field → AliasOpt NAME ArgumentsOpt DirectivesOpt SelectionSetOpt

and writing the rules without resorting to epsilons could become quite
awkward; I don't want to have to come up with a zillion `Field` rules
for all the combinations:

    Field → NAME
    Field → Alias NAME
    Field → NAME Arguments
    Field → Alias NAME Arguments
    Field → NAME Directives
    Field → Alias NAME Directives
    Field → NAME Arguments Directives
    Field → Alias NAME Arguments Directives
    ...
    etc

(That's not even all of them...)

So, this commit is an attempt to get the parser to support epsilon
productions. It does for the small example that I have included in the
tests (`epsilonGrammar`) and does not break the existing grammars which
don't contain epsilon productions.

Having said that, it feels awfully hacky/messy, and I am running
somewhat blind here because there is _so_ much writing out there on the
different forms of LR parsers, the different construction methods etc,
and it is sometimes hard to map what you read (written with specific
assumptions and context) onto what I actually have her (made with
often different assumptions).

However, my attempt to rewrite the fledgling GraphQL grammar to use an
epsilon for `AliasOpt` produces a shift/reduce conflict, which, at first
blush, seems like it might be legit:

    getParseTable(): Shift/Reduce conflict - existing Shift (13) for
    state 7, symbol NAME processing rule: AliasOpt →

At least, this page agrees:

    http://smlweb.cpsc.ucalgary.ca/lalr1.php?grammar=Document+-%3E+DefinitionList+.%0ADefinitionList+-%3E+Definition+.%0ADefinitionList+-%3E+DefinitionList+Definition+.%0ADefinition+-%3E+ExecutableDefinition+.%0AExecutableDefinition+-%3E+OperationDefinition+.%0AOperationDefinition+-%3E+SelectionSet+.%0ASelectionSet+-%3E+open+SelectionList+close+.%0ASelectionList+-%3E+Selection+.%0ASelectionList+-%3E+SelectionList+Selection+.%0ASelection+-%3E+Field+.%0AField+-%3E+AliasOpt+name+.%0AAliasOpt+-%3E+name+colon+%7C+.%0A&substs=