fix(parser): eliminated shift/reduce conflicts
And add a little more to grammar to show it working. We can use optional
productions on the trailing edge of the `Field` rules; it's only in the
leftmost position that having an optional production causes conflicts.

Obviously, keeping the parse table up-to-date in the tests is a bit of
an annoyance. I might switch to snapshot tests to avoid the hassle.