fix(parser): do a better job of handling epsilon productions
Motivated by the comment that you can see being removed in this commit:

> BUG: if I remove directive here, we fail to parse

Some considerable amount of debugging later, of looking at first sets,
follow sets, transition tables, item sets, extended grammars etc, I
figured out what I was doing wrong.

In a nutshell, our extended grammars were busted because the epsilon
productions wound up being unreachable (I was mutating them to encode
the info in their annotations). This in turn meant that the follow sets
were broken (because the rules would never get visited).

We want the rules in the extended grammar, so we break from the
non-extended form (which represents epsilon productions as rules with
empty RHS) and explicitly store such productions with an ε symbol,
annotated just like all the other symbols. I might be able to refactor
this special-casing away in the future, now that I have this working,
but as I was chasing down this fix I just wanted to find a way to
be consistent with the other rules (ie. always having an annotation).

Anyway, having made that decisions, various things flow on from it. We
basically want to reduce using that rule whenever we see such a
production. Tests pass.