refactor(parser): model epsilon productions using empty array
As noted in the last commit, I started modeling epsilon productions in
the extended grammar using an explicit symbol so that I could have
something to annotate. But, having fixed the bug that motivated that
commit in the first place via a circuitous route, I came to the
conclusion that the whole "needing an annotation on the RHS" thing was
probably incidental and not actually necessary.

Sure enough, removing it like I am in this commit leads to a
still-working parser with one less wrinkle in it - now we represent
epsilon productions consistently in both normal and extended grammars.

Comparing the before:

    r17: 6/OperationNameOpt/11 → 11/ε/11

and after[^note]:

    r17: 6/OperationNameOpt/11 → ε

of a rule in a stringified representation of the grammar, we can see why
it's fine to drop the annotation on the RHS: the only actual numbers
that we care about anywhere in there are the ones on the LHS, because
those have to match a rule in order for us to actually look one up and
know how to reduce. Additionally, the initial start state number (the
first number on the LHS) is important because that's the one we use to
key into the `actions` table in order to detect conflicts.

[^note]: One small thing to note here is that the "ε" here is present
only in the stringified representation, to make the epsilon production
visible. In the grammar itself the RHS is a literally empty array of
symbols.