feat(parser): add support for optional default values
This revealed another, clowny bug in `getFollowSets()` 🤦, for which I
am including a fix in this same commit.

Given a rule like:

    VariableDefinition → Variable COLON Type DefaultValueOpt

where `FOLLOW(Type)` included `EQUALS`, I was copying `FOLLOW(Type)`
into `FOLLOW(VariableDefinition)` because `DefaultValueOpt` includes an
epsilon production. This leads to the nonsensical result of `EQUALS`
being in `FOLLOW(VariableDefinition)`.

Instead, we have to do the reverse (copy `FOLLOW(VariableDefinition)`
into `FOLLOW(Type)`). That way, all is rosy with the world.