chore: fix TS lint
Fixes:

    packages/codegen/src/print.ts:199:14 - error TS7030: Not all code paths return a value.

    199         .map((item) => {
                     ~~~~~~~~~~~
Note that there are only two possible values for `item.kind` at this
time, so I could have just used `if`/`else` instead of `if`/`else
if`/`else`. But I like the way the `else if` makes it immediately and
visually obvious that `kind` is `PropertyDeclaration` in the second
branch, so that's why I did it the way I did.