docs(wiki): add some Clipper troubleshooting notes
Specifically for Arch.

I have a bit of a mess at the moment where the `next` branch of
vim-clipper is set up to use some OSC 52 sequence magic when in kitty,
but it doesn't seem to work locally. In the end I went back to the
`main` branch, but that does mean that I still have to figure out how to
get local login sessions and remote sessions via SSH working
consistently still, from both kitty and iTerm2.