test: prime cache before running `yarn test` in CI
I was seeing ts-jest spew a lot of TS errors during `yarn test` in CI:

- https://github.com/wincent/masochist/actions/runs/5766247829/job/15633801387

Of the form:

    ts-jest[ts-compiler] (WARN) packages/lexer/src/RedBlackTree.ts:1:32 - error TS2307: Cannot find module '@masochist/common' or its corresponding type declarations.

    1 import {Queue, invariant} from '@masochist/common';
                                     ~~~~~~~~~~~~~~~~~~~
    packages/lexer/src/RedBlackTree.ts:309:5 - error TS18047: 'h.left' is possibly 'null'.

    309     h.left.color = !h.left.color;
            ~~~~~~

     ... (etc. zillions of type errors in this file)

     PASS  packages/lexer/src/__tests__/RedBlackTree-test.ts

But note, the tests all still pass, and if you re-run `yarn test` they
also pass, but without type errors. The real problem is the first one
— the inability to find the imported module — and all the other
errors after that are just side-effects of running TS with incomplete
information.

Running `yarn run jest --showConfig | grep cache` sheds some light on
why this is happening:

    "cache": true,
    "cacheDirectory": "/private/var/folders/tl/04jnvr4s34v704k5h2rqjmkh0000gn/T/jest_dx",

Before the first `yarn test` run, the cache is empty. By the second run,
it is populated with zillions of files like these; I gather that the
haste-map files probably allow Jest to find the modules on the next run:

    haste-map-e8d74415d3aeb16d2cc8ae21e9fd5b4c-08a35bf6e0c9d8c1c8e54cb80bbda2da-b0edb4f023ba5de968b244d8562f4a59
    jest-transform-cache-0f80fd48a967e2124d2da10822f59505-0bc2a534d3e30a2b12c9a7d83be63585
    perf-cache-0f80fd48a967e2124d2da10822f59505-da39a3ee5e6b4b0d3255bfef95601890

Running `yarn test --no-cache` doesn't fix the problem.

But you _can_ make the first run pass by running a build — that puts
files like `packages/common/tsconfig.tsbuildinfo` and
`packages/common/lib/index.d.ts` etc in place, and ts-jest will make use
of those files and produce a test run free of spurious type errors.

Kind of sucks to have to do a semi-redundant build step in here (given
that we are already verifying `yarn build` in a separate target, but
building is currently quite fast (only a few seconds), so it seems like
a fair trade-off.