fix(lexer): correct bug with character local
As mentioned in previous commit, I had a known bug in there and I wanted
to fix it separately (in this commit) so I could comment on the
trade-off.

I wanted to use `const ch` as an optimization hint to the engine that it
wouldn't have to worry about the type of `ch` changing when it gets
reassigned, and also because immutability is good for avoiding classes
of human-error.

So, my `while` loop optimization used a local `peek` variable to scan
ahead. It would then fall through in many cases to additionally checks,
but these would use the old `ch` variable instead of the current `peek`
one, which meant they were looking at the wrong character even though
`this.index` was right.

With this fix in place I can actually run the benchmarks (best of 3 runs
in each case):

    Before:

        $ node lib/benchmark-static-lexer.js
        Read 5160 bytes
        ┌─────────┬────────────┬───────────┬───────────┬───────────┬──────────────┐
        │ (index) │    rss     │ heapTotal │ heapUsed  │ external  │ arrayBuffers │
        ├─────────┼────────────┼───────────┼───────────┼───────────┼──────────────┤
        │  start  │ '26.94 MB' │ '5.86 MB' │ '5.22 MB' │ '0.35 MB' │  '0.05 MB'   │
        │ warm-up │ '34.74 MB' │ '7.36 MB' │ '5.13 MB' │ '0.32 MB' │  '0.02 MB'   │
        │ finish  │ '34.74 MB' │ '7.36 MB' │ '5.49 MB' │ '0.32 MB' │  '0.02 MB'   │
        └─────────┴────────────┴───────────┴───────────┴───────────┴──────────────┘
        Warm-up: 801.0189800038934ms
        Test: 781.414793998003ms

    After:

        $ node lib/benchmark-static-lexer.js
        Read 5160 bytes
        ┌─────────┬────────────┬───────────┬───────────┬───────────┬──────────────┐
        │ (index) │    rss     │ heapTotal │ heapUsed  │ external  │ arrayBuffers │
        ├─────────┼────────────┼───────────┼───────────┼───────────┼──────────────┤
        │  start  │ '26.89 MB' │ '5.86 MB' │ '5.22 MB' │ '0.35 MB' │  '0.05 MB'   │
        │ warm-up │ '36.40 MB' │ '7.36 MB' │ '5.13 MB' │ '0.32 MB' │  '0.02 MB'   │
        │ finish  │ '36.41 MB' │ '7.36 MB' │ '5.49 MB' │ '0.32 MB' │  '0.02 MB'   │
        └─────────┴────────────┴───────────┴───────────┴───────────┴──────────────┘
        Warm-up: 764.9210250005126ms
        Test: 733.9874399974942ms

    Comparison with reference lexer:

        $ node lib/benchmark-reference-lexer.js
        Read 5160 bytes
        ┌─────────┬────────────┬────────────┬───────────┬───────────┬──────────────┐
        │ (index) │    rss     │ heapTotal  │ heapUsed  │ external  │ arrayBuffers │
        ├─────────┼────────────┼────────────┼───────────┼───────────┼──────────────┤
        │  start  │ '36.70 MB' │ '10.86 MB' │ '6.74 MB' │ '0.44 MB' │  '0.05 MB'   │
        │ warm-up │ '44.21 MB' │ '15.61 MB' │ '7.78 MB' │ '0.42 MB' │  '0.03 MB'   │
        │ finish  │ '52.22 MB' │ '23.61 MB' │ '9.74 MB' │ '0.42 MB' │  '0.03 MB'   │
        └─────────┴────────────┴────────────┴───────────┴───────────┴──────────────┘
        Warm-up: 717.9716149941087ms
        Test: 675.2553099989891ms

ie. we got 6% faster and are now getting closer to the reference
implementation. By implementing some more of the ideas in:

- https://github.com/wincent/masochist/issues/207

we should be able to further close the gap.