fix(lexer): remove erroneous break
I think this was a bug that came in when I moved from the `switch`-based
to an `if`- based state machine implementation.

In the `switch`, the `break` was used to bail out of a `case` block. But
in the `if`, it is bailing out of the outer `while` loop, leading to us
returning a `null` token instead of proceeding and emitting a token
properly.

Wasn't seeing that in my test inputs because the state in question (3)
is reached on seeing a `\r` and my inputs don't have any of those.