feat(codegen): add `walk()`
My kind of change:

    5 files changed, 727 insertions(+), 54 deletions(-)

Replace some ugly code in lex-test.ts with much nicer code. That change
alone is about:

    1 file changed, 47 insertions(+), 40 deletions(-)

(but note, 8 of the added lines are type imports, so the real effect is
about 39 insertions and 40 deletions).

The cost? A 648 `walk()` abstraction that isn't even complete (just
complete enough to fulfill the role I need it for).

Yes, this whole commit message is ironic.