refactor(parser): avoid `$$` local
Doesn't make it faster, but does make it look cleaner.

Might want to document this as a requirement for semantic actions (that
they should be written such that `$$ =` can be safely substituted by
`return`.

But more likely I can do some static analysis to verify that that _is_
the case before doing it.