feat(typescript): start making simple parser for TypeScript (subset)
Definitely only a subset. Idea being to replace adhoc parsing currently
in `ast.ts` with a little mini parser that is enough to parse my TS
templates and probably some semantic actions as well.

There'll be a chicken-and-egg thing going on here in the end (in the
sense that the lexer and parser rely on the codegen module, and the
codegen module is going to rely on the TS parser), but because I can
check in the CIGAR, I can use that to bootstrap one or the other.