test: avoid duplicate snapshots
Bun currently doesn't cope well with some of the test in this snapshot.
Specifically, it hates this line:

    throw new Error(\`Unsupported operation type: ${contents}\`);

Every time you run `bun test`, it adds a duplicate copy of the snapshot
to the file.

At first I thought it was caused by a backtick string with interpolation
in it, but it's actually just the interpolation that matters, because
the whole thing is in backticks. ie.

    exports[`some name`] = `
      ...
      \`stuff ${something} \`
      ...
    `;

So, the "${contents}" substring is being treated as an interpolation in
the outer string, and of course, there is no `contents`. I can see why
this would cause mismatches, but I'm not actually sure why it causes
duplicates. In any case, a bug has been filed with a minimal repro.

See: https://github.com/oven-sh/bun/issues/4722