feat: expand TypeScript statement parser
Required me to extend the lexer.

Which didn't work at first, so I added the "dotify" script to help me
debug. The images it produces are almost unusably large, though, even
though the state machine doesn't have much more than 100 states.

Includes gratuitous fix to "build.ts" in the parser generator to ensue
it generates:

    return $1;

instead of:

    return$1;

in the event that the semantic action is written as:

   { $$ =$1; }

instead of:

   { $$ = $1; }

(discovered that weakness because one of my actions was originally like
that.)