refactor: inline Token class into generated lexers
This is needed to avoid a circular dependency:

- "lexer" + "parser" need "codegen" to build generated lexers/parsers
- "typescript" + "graphql" use "lexer" + "parser"
- "codegen" needs parser/lexer from "typescript"

So, the plan is to _copy_ the Typescript parser into "codegen" at build
time. Having the generated lexer (and parser) depend on `Token` means
that the copying still leaves us with a cycle:

- "codegen" will depend on "lexer" for `Token`

ie. "codegen" -> "lexer" -> "codegen"

So, we break that cycle by copying `Token`. This is ok because TS uses
structural instead of nominal typing (ie. all `Token` classes are equal,
so we don't have to change our tests), because the duplication is
non-fragile (we copy the file contents programmatically instead of by
hand) and will become less fragile in the future (when I can parse the
Token implementation into an AST instead of using `RawStatement`).