feat(parser): lay groundwork for operator precedence/associativity
Added "todo" test to show why this is necessary. I am not sure if I am
doing something wrong or not, but the way I've written the
`MemberExpression` production, I'm getting a right-associative AST node
when nesting, when it should be left-associative.

In any case, it seems useful to be able to specify precedence, and I can
see myself needing it as I grow the TS grammars, so let's get started.

Following the pattern laid out in the Bison docs:

- http://web.mit.edu/gnu/doc/html/bison_8.html
- https://www.gnu.org/software/bison/manual/bison.html#How-Precedence

Namely:

> How Precedence Works
>
> The first effect of the precedence declarations is to assign
> precedence levels to the terminal symbols declared. The second effect
> is to assign precedence levels to certain rules: each rule gets its
> precedence from the last terminal symbol mentioned in the components.
>
> Finally, the resolution of conflicts works by comparing the precedence
> of the rule being considered with that of the look-ahead token. If the
> token's precedence is higher, the choice is to shift. If the rule's
> precedence is higher, the choice is to reduce. If they have equal
> precedence, the choice is made based on the associativity of that
> precedence level.
>
> Not all rules and not all tokens have precedence. If either the rule
> or the look-ahead token has no precedence, then the default is to
> shift.

Some teasing out is required to get what is meant by "If they have
equal precedence, the choice is made based on the associativity of that
precedence level". The answer to that is spelled out higher up in the
docs, which say:

> What about input such as `1 - 2 - 5'; should this be `(1 - 2) - 5' or
> should it be `1 - (2 - 5)'? For most operators we prefer the former,
> which is called left association. The latter alternative, right
> association, is desirable for assignment operators. The choice of
> left or right association is a matter of whether the parser chooses
> to shift or reduce when the stack contains `1 - 2' and the look-ahead
> token is `-': shifting makes right-associativity.

I did some searching in other places before landing on the Bison docs. I
think the latter do the best job of explaining what to do, but there are
other takes I found along the way:

- https://stackoverflow.com/a/33559685/2103996
- https://stackoverflow.com/a/48605074/2103996
- https://stackoverflow.com/a/12417762/2103996