feat: parallel formatting (demo, will revert)
The overhead of all these slow async things (async generator, async `fs`
calls, worker communication etc) make this not worth. No faster than
running in process, serially.

Committing so I can get it in the history and revert in the next commit.
It's not even finished (implemented checking but not formatting). Just
wanted to test this out out of curiosity.