perf: switch to dprint for formatting
Tried to pick settings that would minimize churn, but the churn that
remains is worth it IMO because now `bun format` and `bun format:check`
run in about 0.7s when the cache is warm (previously, they took about 4
seconds).

Some things dprint does that I don't like (like spaces inside braces for
objects that fit on one-line in JSON, or some overzealous
line-breaking), but speed matters, so I'm going with it.

Closes: https://twitter.com/wincent/status/1705567589391270240