feat(codegen): make it possible to make typed production functions
Sneaking in: minor refactoring of the grammar (reusing ExpressionList
because there is no need for redundant ArrayItemList).

With this, the entire 7K+ line "parseStatement.ts" typechecks correctly,
with only a few small errors to deal with in the final function at the
end, which will be easily dealt with in a future commit.

As noted in the comment, my only concern is with the self-import that
might cause `tsc` to get confused.